T: move tests for duplicate definition to separate class
